Fix: no more @token_required
